Added ability to cycle through rooms with LMB/RMB (#28)
* Added ability to cycle through rooms with LMB/RMB

Simply "increments" or "decrements" the room ID (i.e. if room ID is "005", it will go to "004" or "006")
Had to make map public in Game1.cs, and currentRoom public in Map.cs

* Fixed door display

Made the map dictionary and currentAdjacentList public so that they could be accessed from the MouseController class, then updated the map dictionary so that it would be able to get the doors needed.