Fixed update, added more logic, added sound
Had to set player's health to -1 in Game1.cs to avoid the Die() function being called over and over at 0 health. Die() function now stops the music to allow for the dying sound to play. SoundStorage's music_instance field had to be made public to do this. Had to get creative with the "pop" animation followed by drawing nothing for when Link has fully died, as can be seen in the LinkTextureStorage class. Everything should be working from here, albeit with some sloppy logic in Update.